FlappyBirdCopy Reusable Repo
Reusable Repo that serves as Distribution Repo, so all assets from here are distributed to BasePattern that then uses these both for building and for distribution to Vercel. Vercel in turn uses Assets as the folder name, so now we have solved the location alignment problem: We only need one location path: 'assets'.